requirements.txt
langchain
langchain-community
langchain-core
langchain-chroma
langchain-groq
langchain-huggingface
langchain-text-splitters
python-dotenv
streamlit
unstructured
nltk
chromadb
sentence-transformers
pypdf